Removed StringEntry nested class from Queue in attempt to have loose coupling. StringEntry and Queue are now apart of the QueueNS namespace.
